+++ global.R
@@ -5,65 +5,68 @@ files.df <- NULL
 SHEETS <- NULL
 NAMES <- NULL
 
-# Preprocess QS results from Excel
-processQSResultsExcel <- function(raw) {
-  # Extract variables
-  # - Experiment Name
-  # - Experiment Run End Time
-  # - Instrument Type
-  # - Instrument Serial Number
-  exp.df <- raw[which(raw$`Block Type` %in% c("Experiment Name", "Experiment Run End Time", "Instrument Type", "Instrument Serial Number")), 1:2]
-  exp.name <- as.character(exp.df[1, 2])
-  if(nchar(as.character(exp.df[2, 2])) == "24") {
-    exp.date <- substr(as.character(exp.df[2, 2]), 1, 16)
+readTxt <- function(file.dir) {
+  # Tab file contains three parts
+  # 1. run settings and info
+  # 2. results
+  # 3. melt curve results
+  
+  # First read file and check the number 
+  con = file(file.dir, "r")
+  i = 1
+  while(TRUE) {
+    line = readLines(con, n = 1)
+    
+    # Extract variables
+    # - Experiment Run End Time
+    # - Instrument Type
+    # - Instrument Serial Number
+    
+    if(substr(line, 1, 28) == "* Experiment Run End Time = ") {
+      exp.date <- substr(line, 29, nchar(line))
+      #print(exp.date)
+    }
+    
+    if(substr(line, 1, 26) == "* Instrument Name =       ") {
+      exp.instr <- substr(line, 27, nchar(line))
+      #print(exp.instr)
+    }
+    
+    if(substr(line, 1, 29) == "* Instrument Serial Number = ") {
+      exp.instr.id <- substr(line, 30, nchar(line))
+      #print(exp.instr.id)
+    }
+    
+    if(line == "[Results]") {
+      break
+    }
+    i = i + 1
+  }
+  j = 0
+  while(TRUE) {
+    line = readLines(con, n = 1)
+    if(line == "") {
+      break
+    }
+    j = j + 1
+  }
+  close(con)
+  
+  # Read file
+  r.df <- suppressMessages(readr::read_tsv(file.dir, skip = i, n_max = j-1))
+  
+  # Fix date
+  print(exp.date)
+  if(nchar(as.character(exp.date)) == 24) {
+    exp.date <- substr(as.character(exp.date), 1, 16)
   } else {
-    exp.date <- format(as.POSIXct(substr(as.character(exp.df[2, 2]), 1, 16)), "%d-%m-%Y %H:%M")
-  }
-  exp.instr <- as.character(exp.df[3, 2])
-  exp.instr.id <- as.character(exp.df[4, 2])
-  
-  # Preprocess results
-  idx <- which(raw$`Block Type` == "Well")
-  r.df <- raw[(idx+1):nrow(raw), ]
-  colnames(r.df) <- unname(unlist(lapply(raw[(idx), , drop = T], function(x) as.character(x[[1]]))))
-  rownames(r.df) <- NULL
-  
-  # Check if MTP exists
-  if(!any(colnames(r.df) %in% "MTP")) {
-    r.df <- r.df %>% mutate(MTP = "N")
+    exp.date <- format(as.POSIXct(substr(as.character(exp.date), 1, 16)), "%d-%m-%Y %H:%M")
   }
   
-  r.df$MTP <- factor(r.df$MTP, levels = c("N", "Y"))
-  
-  # Filter relevant columns
-  r.df <- r.df[, c("Sample Name", "Target Name", "CT", "Cq Conf", "MTP", "Tm1")]
-  
-  # Rename columns
-  r.df <- r.df %>% 
-    rename(Ct = CT,
-           `Sample ID` = `Sample Name`)
-  
-  # Order gene names
-  r.df <- r.df %>% 
-    mutate(`Target Name` = factor(r.df$`Target Name`,
-                                  levels = unique(r.df$`Target Name`), 
-                                  ordered = T))
-  
-  # Reclass columns
-  r.df <- r.df %>% 
-    mutate(`Sample ID` = as.character(`Sample ID`),
-           Ct = as.numeric(as.character(Ct)),
-           `Cq Conf` = as.numeric(as.character(`Cq Conf`)),
-           MTP = as.character(MTP),
-           Tm1 = as.numeric(as.character(Tm1))) %>% 
-    as_tibble()
-  
-  # Round columns with numeric values
-  r.df <- r.df %>% mutate_if(is.numeric, function(x) round(x+100*.Machine$double.eps, 3))
-  
+  print(exp.date)
+  print(class(exp.date))
   # Compile output as list object
   obj <- list(data = r.df,
-              name = exp.name,
               date = exp.date,
               instr = exp.instr,
               id = exp.instr.id)
@@ -107,42 +110,32 @@ processQSResultsTxt <- function(r.df) {
   return(r.df)
 }
 
-preProcessFiles <- function(files.df) {
+preProcessFiles <- function(if.df) {
   # 4. Validate and preprocess selected files
   df <- data.frame()
   
-  for(i in 1:length(files.df$name)) {
-    file.name <- files.df$name[i]
-    path <- files.df$datapath[i]
+  for(i in 1:nrow(if.df)) {
+    path <- if.df$datapath[i]
     
-    if(substr(path, nchar(path)-3, nchar(path)) == "xlsx") {
-      raw <- suppressMessages(readxl::read_xlsx(path = as.character(path),
-                                                sheet = "Results"))
-      pp <- processQSResultsExcel(raw)
-    } else if(substr(path, nchar(path)-2, nchar(path)) == "txt") {
-      pp <- readTxt(path)
-      pp$data <- processQSResultsTxt(pp$data)
-    }
+    pp <- readTxt(path)
+    
+    pp$data <- processQSResultsTxt(pp$data)
     
     pp <- pp$data %>% 
       mutate(run = i,
-             ID = pp$name,
+             `Exp name` = if.df$`Exp name`[i],
              date = pp$date,
              instrument = pp$instr,
              id = pp$id) %>% 
-      select(ID, instrument, date, everything())
+      select(`Exp name`, instrument, date, everything())
     
     df <- rbind(df, pp)
   }
+  
   return(df)
 }
 
-prepareDataXlsx <- function(df) {
-  df <- df %>% 
-    mutate(date = as.POSIXct(date, format = "%d-%m-%Y %H:%M")) %>% 
-    #arrange(date) %>% 
-    mutate(date = as.character(date))
-  
+prepareOutputDataXlsx <- function(df) {
   Ct <- data.frame()
   `Cq Conf` <- data.frame()
   pc <- data.frame()
@@ -157,27 +150,27 @@ prepareDataXlsx <- function(df) {
     temp.df$`Sample ID` <- factor(temp.df$`Sample ID`, levels = unique(temp.df$`Sample ID`), ordered = T)
     
     Ct.raw <- temp.df %>% 
-      select(date, ID, `Sample ID`, `Target Name`, Ct) %>%
+      select(date, `Exp name`, `Sample ID`, `Target Name`, Ct) %>%
       spread(`Target Name`, Ct) %>%
       filter(`Sample ID` != "Positive Control" & `Sample ID` != "Negative Control") 
     `Cq Conf.raw` <- temp.df %>% 
-      select(date, ID, `Sample ID`, `Target Name`, `Cq Conf`) %>%
+      select(date, `Exp name`, `Sample ID`, `Target Name`, `Cq Conf`) %>%
       spread(`Target Name`, `Cq Conf`) %>% 
       filter(`Sample ID` != "Positive Control" & `Sample ID` != "Negative Control")
     pc.raw <- temp.df %>%
-      select(date, ID, `Sample ID`, `Target Name`, Ct) %>%
+      select(date, `Exp name`, `Sample ID`, `Target Name`, Ct) %>%
       spread(`Target Name`, Ct) %>%
       filter(`Sample ID` == "Positive Control")
     nc.raw <- temp.df %>% 
-      select(date, ID, `Sample ID`, `Target Name`, Ct) %>%
+      select(date, `Exp name`, `Sample ID`, `Target Name`, Ct) %>%
       spread(`Target Name`, Ct) %>%
       filter(`Sample ID` == "Negative Control")
     MTP.raw <- temp.df %>% 
-      select(date, ID, `Sample ID`, `Target Name`, MTP) %>%
+      select(date, `Exp name`, `Sample ID`, `Target Name`, MTP) %>%
       spread(`Target Name`, MTP) %>% 
       filter(`Sample ID` != "Positive Control" & `Sample ID` != "Negative Control")
     Tm1.raw <- temp.df %>% 
-      select(date, ID, `Sample ID`, `Target Name`, Tm1) %>%
+      select(date, `Exp name`, `Sample ID`, `Target Name`, Tm1) %>%
       spread(`Target Name`, Tm1) %>% 
       filter(`Sample ID` != "Positive Control" & `Sample ID` != "Negative Control")
     
@@ -223,75 +216,3 @@ x <- c("29 79 110",
 colors <- sapply(strsplit(x, " "), function(x)
   rgb(x[1], x[2], x[3], maxColorValue=255))
 
-readTxt <- function(file.dir) {
-  # Tab file contains three parts
-  # 1. run settings and info
-  # 2. results
-  # 3. melt curve results
-  
-  # First read file and check the number 
-  con = file(file.dir, "r")
-  i = 1
-  while(TRUE) {
-    line = readLines(con, n = 1)
-    
-    # Extract variables
-    # - Experiment Name
-    # - Experiment Run End Time
-    # - Instrument Type
-    # - Instrument Serial Number
-    
-    if(substr(line, 1, 20) == "* Experiment Name = ") {
-      exp.name <- substr(line, 21, nchar(line))
-      print(exp.name)
-    }
-    
-    if(substr(line, 1, 28) == "* Experiment Run End Time = ") {
-      exp.date <- substr(line, 29, nchar(line))
-      print(exp.date)
-    }
-    
-    if(substr(line, 1, 26) == "* Instrument Name =       ") {
-      exp.instr <- substr(line, 27, nchar(line))
-      print(exp.instr)
-    }
-    
-    if(substr(line, 1, 29) == "* Instrument Serial Number = ") {
-      exp.instr.id <- substr(line, 30, nchar(line))
-      print(exp.instr.id)
-    }
-    
-    if(line == "[Results]") {
-      break
-    }
-    i = i + 1
-  }
-  j = 0
-  while(TRUE) {
-    line = readLines(con, n = 1)
-    if(line == "") {
-      break
-    }
-    j = j + 1
-  }
-  close(con)
-  
-  # Read file
-  r.df <- readr::read_tsv(file.dir, skip = i, n_max = j-1)
-  
-  # Fix date
-  if(nchar(as.character(exp.date) == "24")) {
-    exp.date <- substr(as.character(exp.date), 1, 16)
-  } else {
-    exp.date <- format(as.POSIXct(substr(as.character(exp.date), 1, 16)), "%d-%m-%Y %H:%M")
-  }
-  
-  # Compile output as list object
-  obj <- list(data = r.df,
-              name = exp.name,
-              date = exp.date,
-              instr = exp.instr,
-              id = exp.instr.id)
-  
-  return(obj)
-}